+++ YAPM/Providers/Classes/Handles/cHandle.vb
@@ -83,7 +83,7 @@ Public Class cHandle
             asyncCallbackHandleUnload.poolObj(Me.Infos.ProcessID, Me.Infos.Handle, newAction))
 
     End Function
-    Private Sub unloadHandleDone(ByVal Success As Boolean, ByVal pid As Integer, ByVal handle As Integer, ByVal msg As String, ByVal actionNumber As Integer)
+    Private Sub unloadHandleDone(ByVal Success As Boolean, ByVal pid As Integer, ByVal handle As IntPtr, ByVal msg As String, ByVal actionNumber As Integer)
         If Success = False Then
             MsgBox("Error : " & msg, MsgBoxStyle.Exclamation Or MsgBoxStyle.OkOnly, _
                    "Could not unload handle " & handle.ToString)
@@ -233,7 +233,7 @@ Public Class cHandle
 
         AddSharedPendingTask(newAction, t)
     End Function
-    Private Shared Sub unloadsharedHandleDone(ByVal Success As Boolean, ByVal pid As Integer, ByVal handle As Integer, ByVal msg As String, ByVal actionNumber As Integer)
+    Private Shared Sub unloadsharedHandleDone(ByVal Success As Boolean, ByVal pid As Integer, ByVal handle As IntPtr, ByVal msg As String, ByVal actionNumber As Integer)
         If Success = False Then
             MsgBox("Error : " & msg, MsgBoxStyle.Exclamation Or MsgBoxStyle.OkOnly, _
                    "Could not unload handle " & handle.ToString)
